% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthSimulation.R
\name{add.compounds}
\alias{add.compounds}
\title{Add compounds to the growth environment}
\usage{
add.compounds(object, compounds, concentrations)
}
\arguments{
\item{object}{S4-object of type \code{growthSimulation}.}

\item{compounds}{Character vector with the compound IDs of substances to add to the environment. Compound IDs should correspond to the models' exchange reactions IDs ("EX_[cpdid]"), without the "EX_" prefix.}

\item{concentrations}{Numeric vector with the concentrations of the compounds from \code{compounds} in the same order. Values in mM.}
}
\value{
Return a S4-object of type \code{growthSimulation}.
}
\description{
The function can be used to add substances to the growth environment by providing concentrations in mM.
}
\details{
Compound concentration are equally distributed across the whole growth environment. More options are planned.
}
\examples{
sim <- init.simulation(cbind(c(-100, -100, 100, 100), c(-100, 100, 100, -100)),
                       gridFieldSize = 2, gridFieldLayers = 5)

sim <- add.compounds(sim,
                     compounds = c("cpd00027_e0","cpd00029_e0","cpd00047_e0",
                                   "cpd00159_e0","cpd00211_e0"),
                     concentrations = c(50,0,0,0,0))

}
