% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exoenzyme.R
\name{add.exoenzyme}
\alias{add.exoenzyme}
\title{Add an exoenzyme to organism and simulation}
\usage{
add.exoenzyme(
  object,
  organism,
  id,
  mets,
  stoich,
  production.rate = 0.01,
  name = NULL,
  D = 10,
  lambda = 0.4,
  Kcat = 1000,
  Km = 25,
  init.conc = 0
)
}
\arguments{
\item{object}{S4-object of type \link{growthSimulation}}

\item{organism}{Character indicating the name of the organism, to which the
exoenzyme production is linked}

\item{id}{Character defining the ID used for the specific exoenzyme}

\item{mets}{Character vector with the compound IDs that participate in reaction
catalyzed by the exoenzyme.}

\item{stoich}{Numeric vector of the same length and order as `mets` specifying
the stoichiometries of compounds. First entry (the substrate's coefficient)
should be -1.}

\item{production.rate}{Numeric indicating the production rate of the enzyme
by the organism. Unit: mg Enzyme / gDW cell. Default: 0.01}

\item{name}{Character with an optional name for the enzyme.}

\item{D}{Diffusion coefficient of the enzyme. Unit µm^2/s. Dedault: 10}

\item{lambda}{Numeric indicating the decay rate of the enzyme. Unit: per hour.
Default: 0.4 . The enzyme's half life can be calculated by ln(2)/lambda}

\item{Kcat}{Numeric for enzyme's turnover rate. Unit: 1/s . Default: 1000}

\item{Km}{Numeric for Menten-Michaelis Km value. Unit: mM . Default: 25}

\item{init.conc}{Numeric indicating the initial concentration of the enzyme
in the growth environment. Unit: nM . Default: 0}
}
\description{
Adds a new exoenzyme to the simulation and links it to a specific
organism.
}
