% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exoenzyme.R
\name{add_exoenzyme}
\alias{add_exoenzyme}
\title{Add an exoenzyme to organism and simulation}
\usage{
add_exoenzyme(
  object,
  organism,
  id,
  mets,
  stoich,
  production.rate = 0.01,
  name = NULL,
  D = 10,
  lambda = 0.4,
  Kcat = 10000,
  Km = 100,
  init.conc = 0
)
}
\arguments{
\item{object}{S4-object of type \link{growthSimulation}}

\item{organism}{Character indicating the name of the organism, to which the
exoenzyme production is linked}

\item{id}{Character defining the ID used for the specific exoenzyme}

\item{mets}{Character vector with the compound IDs that participate in reaction
catalyzed by the exoenzyme.}

\item{stoich}{Numeric vector of the same length and order as `mets` specifying
the stoichiometries of compounds. First entry (the substrate's coefficient)
should be -1.}

\item{production.rate}{Numeric indicating the production rate of the enzyme
by the organism. Unit: nmol enzyme catalytic centers per gDW cells per hr.
Default: 0.01}

\item{name}{Character with an optional name for the enzyme.}

\item{D}{Diffusion coefficient of the enzyme. Unit µm^2/s. Default: 10}

\item{lambda}{Numeric indicating the decay rate of the enzyme. Unit: per hour.
Default: 0.4 . The enzyme's half life can be calculated by ln(2)/lambda}

\item{Kcat}{Numeric for enzyme's turnover rate. Unit: 1/s . Default: 10000}

\item{Km}{Numeric for Menten-Michaelis Km value. Unit: mM . Default: 100}

\item{init.conc}{Numeric indicating the initial concentration of the enzyme
in the growth environment. Unit: nM . Default: 0}
}
\description{
Adds a new exoenzyme to the simulation and links it to a specific
organism.
}
\details{
Exoenzymes can differ markedly in their kinetic parameters. The defaults
provided here do not represent a typical enzymes, but are chosen based
on data from inveratases (EC 3.2.1.26) from Zymomonas mobilis. Kcat (~10000 s^-1)
and Km (100 mM) were obtained from https://www.brenda-enzymes.org/enzyme.php?ecno=3.2.1.26&Suchword=&reference=&UniProtAcc=&organism%5B%5D=Zymomonas+mobilis&show_tm=0.
}
