% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.uptake.R
\name{adjust.uptake,modelorg,numeric,list,numeric-method}
\alias{adjust.uptake,modelorg,numeric,list,numeric-method}
\title{Adjust a model's lower bounds for exchange reactions based on compound accessibility}
\usage{
\S4method{adjust.uptake}{modelorg,numeric,list,numeric}(model, cMass, accCpdFMOL, deltaTime)
}
\description{
Adjust a model's lower bounds for exchange reactions based on compound accessibility
}
