% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthSimulation.R
\name{dilute_compounds}
\alias{dilute_compounds}
\title{Dilute compounds}
\usage{
dilute_compounds(
  object,
  dilution.factor,
  compounds = NULL,
  incl.constant = FALSE
)
}
\arguments{
\item{object}{A \link{growthSimulation} object}

\item{dilution.factor}{Numeric within the range of [0,1], by which the
compound concentrations are diluted. `1` completely dilutes concentrations to
0 mM, while `0` does not change anything.}

\item{compounds}{Character of compound IDs that should be diluted. If `NULL`,
all compounds are diluted.}

\item{incl.constant}{Logical specifying whether also constant compounds
should be diluted. Default: FALSE}
}
\description{
Dilutes all or selected compounds with a given dilution factor.
}
