% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EcoAgents-package.R
\docType{package}
\name{EcoAgents-package}
\alias{EcoAgents-package}
\alias{EcoAgents}
\title{EcoAgents}
\description{
EcoAgents: Agent-based modelling of microbial communities in time and continuous µ-meter-scale space
}
