% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exoenzyme.R
\docType{class}
\name{Exoenzyme-class}
\alias{Exoenzyme-class}
\alias{Exoenzyme}
\title{Structure of the S4 class "Exoenzyme"}
\description{
Structure of the S4 class "Exoenzyme"
}
\section{Slots}{

\describe{
\item{\code{id}}{Character for exoenzyme ID}

\item{\code{name}}{Characer of exoenzyme name}

\item{\code{D}}{Numeric of the enzymes' diffusion coefficients. Unit: \eqn{\mu}m^2 per
sec}

\item{\code{lambda}}{Numeric for enzyme's decay constants. Unit: per hr}

\item{\code{Kcat}}{Numeric for turnover rate (kcat) in MM-kinetics. Unit: 1/s}

\item{\code{Km}}{Numeric for Km in MM-kinetics. Unit: mM}

\item{\code{mets}}{Character vector of metabolite IDs for compounds in the reaction.
Main substrate should be the first element.}

\item{\code{stoich}}{Numeric vector with the stoichiometries of `mets` in exoenzyme's
reaction.}
}}

