% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthEnvironment.R
\docType{class}
\name{growthEnvironment-class}
\alias{growthEnvironment-class}
\alias{growthEnvironment}
\title{Structure of the S4 class "growthEnvironment"}
\description{
Structure of the S4 class "growthEnvironment"
}
\section{Slots}{

\describe{
\item{\code{field.pts}}{Object of class \link[sp]{SpatialPoints}. Coordinates of
rhombic dodecahedron centers points.}

\item{\code{compounds}}{Character vector with compound IDs}

\item{\code{compound.names}}{Character vector with compound names}

\item{\code{compound.D}}{Numeric vector of the diffusion coefficient values for each
compound. Unit: \eqn{\mu}m^2/s}

\item{\code{concentrations}}{(n x m) numeric matrix with n columns representing compounds
and m grid field (field.pts). Units of matrix entries: mM}

\item{\code{conc.isConstant}}{Logical vector indicating if the respective compound is
fixed (i.e. buffered) in its concentration.}

\item{\code{fieldLayers}}{Integer specifying how many layers (z-dimension) of rhomic
dodecahedra fields are in the environment representation.}

\item{\code{nfields}}{Integer indicating the number of fields in the environment
representation.}

\item{\code{fieldSize}}{Size of each rhomic dodecahedron in \eqn{\mu}m as the distance between
opposite faces.}

\item{\code{fieldVol}}{Volume of each field. Unit: \eqn{\mu}m^3}

\item{\code{mat.in}}{A two column numeric matrix as a representation of the grid environment
as a directed graph. First column (From) and second column (To) are denoting
field's indices in `field.pts`}

\item{\code{mat.out}}{A two column matrix. First column: indices of each field in `field.pts`;
second column: Number of neighboring fields}

\item{\code{exoenzymes}}{A list with the \link{Exoenzyme} S4 objects.}

\item{\code{exoenzymes.conc}}{Same as `concentrations`, but for the exoenzymes and in
nM.}
}}

