% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Organism.R
\docType{class}
\name{Organism-class}
\alias{Organism-class}
\alias{Organism}
\title{Structure of the S4 class "Organism"}
\description{
Structure of the S4 class "Organism"
}
\section{Slots}{

\describe{
\item{\code{cellDiameter}}{Numeric for the initial Diameter of cell spheres in µm}

\item{\code{cellMassInit}}{Numeric for the initial cell mass in pg}

\item{\code{cellMassAtDivision}}{Limit of a cell's mass before it divided into two
daughter cells. Unit: pg}

\item{\code{cellShape}}{Character for the cell shape type. Currently, only coccus/sphere
shapes are supported.}

\item{\code{vmax}}{Numeric for the maximum speed a cell can move in µm/s}

\item{\code{scavangeDist}}{Numeric indicating the maximum distance (from cell surface)
a cell can scavenge nutrients from its surrounding. Unit: µm}

\item{\code{chemotaxisCompound}}{Character vector with the compound IDs that influence
the cells chemotaxis behavior.}

\item{\code{chemotaxisStrength}}{Numeric vector indicating the strength of chemotaxis.
Positive: attracting; Negative: Repelling}

\item{\code{mod}}{Object of S4-class \link[sybil]{modelorg} for the organisms metabolic
network model.}

\item{\code{exoenzymes}}{Character vector of IDs of the organism's exoenzymes}

\item{\code{exoenzymes.prod}}{Numeric vector of the production rates of exoenzymes.
Unit: nmol / gDW / hr (nmol Enzyme per gDW cells per hr)}
}}

