% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthSimulation.R
\name{plot.compounds}
\alias{plot.compounds}
\title{Plot compound time curves}
\usage{
plot.compounds(object, compounds = NULL, tlim = NULL, ylim = NULL, ...)
}
\arguments{
\item{object}{S4-object of type \code{growthSimulation}.}

\item{compounds}{Vector of compound IDs whose concentration is plotted. If no IDs are provided,
the top variable (SD) compounds are plotted (max. 8 compounds).}

\item{tlim}{Numeric vector of length 2, specifying the x-range (Time) to be displayed.}

\item{ylim}{Numeric vector of length 2, specifying the y-range (mM) to be displayed.}
}
\value{
A ggplot object.
}
\description{
Plot concentrations of metabolites over time
}
