% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot.environment.exoenzymes,growthSimulation,character-method}
\alias{plot.environment.exoenzymes,growthSimulation,character-method}
\alias{plot.environment.exoenzymes}
\title{Plot spatial distribution of exoenzymes}
\usage{
\S4method{plot.environment.exoenzymes}{growthSimulation,character}(
  object,
  exoenzymes,
  exoenzyme.names = NULL,
  xlim = NULL,
  ylim = NULL,
  iter = NULL,
  scalebar.color = "white",
  layer = 0,
  gradient.limits = NULL,
  gradient.option = "viridis"
)
}
\arguments{
\item{object}{S4-object of type \link{growthSimulation}}

\item{exoenzymes}{Character string of exoenzyme ids to plot}

\item{exoenzyme.names}{Character string of exoenzyme names that should be
displayed as facet header instead of exoenzyme names from the simulation
object.}

\item{xlim}{Numeric vector of length 2 specifying the limits (left and right)
for x axis; i.e. the horizontal dimension.}

\item{ylim}{Numeric vector of length 2 specifying the limits (top and bottom)
for y axis; i.e. the vertical dimension.}

\item{iter}{Positive integer number of the simulation step/iteration to plot
the distribution. Works only if the respective metabolite
concentrations were prior recorded (see \link{run.simulation}). If
\code{NULL}, current distribution of exoenzyme concentrations.}

\item{scalebar.color}{Color of the scale bar and its annotation. Default:
"white".}

\item{layer}{Positive integer, specifying which layer (z-dimension) to plot.
Default: 0 (base plane).}

\item{gradient.limits}{Numeric vector of length 2 specifying the
concentration range that is spanned by the color gradient.}

\item{gradient.option}{Character string indicating the colormap to be used
for visualizing exoenzyme concentrations. Please refer to \link[ggplot2]{scale_colour_viridis_d}
so see possible options.}
}
\value{
A \link{ggplot}
}
\description{
Plots the spatial distribution of exoenzyme concentrations in a
heatmap.
}
