% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scavenge.compounds2.R
\name{scavenge.compounds,list,matrix,character,numeric-method}
\alias{scavenge.compounds,list,matrix,character,numeric-method}
\title{Scavenge available compounds from local environment}
\usage{
\S4method{scavenge.compounds}{list,matrix,character,numeric}(localEnv, env_conc, env_cpds, env_fieldVol)
}
\arguments{
\item{localEnv}{data.table for local environment. The expected columns elements are: \code{field.id} is a
vector if field indeces of the environment grid; \code{field.dist} the distance to each of the respective fields
within the local environment; \code{acc.prop} specifies the accessible proportion of the field to the cell.}

\item{object}{A \code{growthSimulation} object.}
}
\value{
Lists with the following elements: \code{compounds}: character vector of compound ids, and \code{fmol}: numeric
vector with the absolute metabolite availability in fmol.
}
\description{
Scavenge available compounds from local environment
}
