% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthSimulation.R
\name{show,growthSimulation-method}
\alias{show,growthSimulation-method}
\title{Print a short summary of a growth simulation}
\usage{
\S4method{show}{growthSimulation}(object)
}
\arguments{
\item{object}{S4-object of type \link{growthSimulation}.}
}
\description{
Displays a few numbers to describe the current status of a growth simulation.
}
