% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthSimulation.R
\name{summary.exchanges}
\alias{summary.exchanges}
\title{Summary of uptake / production by organism type}
\usage{
summary.exchanges(object, iter = NULL)
}
\arguments{
\item{object}{S4-object of type \code{growthSimulation}.}

\item{iter}{Positive integer number of the simulation step/iteration to plot the rates.}
}
\value{
A data.table.
}
\description{
Uptake/Production rates in fmol summarized by organism type
}
