% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthSimulation.R
\docType{class}
\name{growthSimulation-class}
\alias{growthSimulation-class}
\title{Structure of the S4 class "growthSimulation"}
\description{
Structure of the S4 class \code{growthSimulation} as framework for the grwoth environment and container for agent-based flux balance analysis.
}
\section{Slots}{

\describe{
\item{\code{n_rounds}}{integer for the number of simulation rounds that were already performed for this \code{growthSimulation} object.}

\item{\code{deltaTime}}{double for length of each time step for the simulation in hours.}

\item{\code{diffusionNIter}}{integer. Number of diffusion steps per simulation round.}

\item{\code{rMotion}}{double. Maximum x- and y distance a cell can travel by means of random movement per simulation round.}

\item{\code{models}}{list for \code{Organism} objects to represent the different strains in the simulation.}

\item{\code{history}}{list with recordings of simulation status information at each simulation round.}

\item{\code{universePolygon}}{Matrix specifiying the corners of the polygon that defined the growth environment boundaries. 2-dimensional: x and y.}

\item{\code{environ}}{Object of S4-class \code{grwothEnvironment}, that specifies the environment mesh layout, compounds, and their concentrations.}
}}

