% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthSimulation.R
\name{init.simulation}
\alias{init.simulation}
\title{Initialise a growth simulation.}
\usage{
init.simulation(
  universePolygon,
  gridFieldSize = 1,
  gridFieldLayers = 3,
  deltaTime = 1/6,
  pFBAcoeff = 1e-06,
  rMotion = 0.1,
  ...
)
}
\arguments{
\item{universePolygon}{A two column matrix specifiying the x and y coordinates of the polygon, that describes the growth environment boundaries.}

\item{gridFieldSize}{double. Distance between neighboring environments 3D mesh field elements (rhombic dodecahedrons) in µm.}

\item{gridFieldLayers}{integer. z-dimension (height) as the number of layers of field elements.}

\item{deltaTime}{double specifying the length of each time step for the simulation in hours.}

\item{pFBAcoeff}{integer. pFBA coefficient. Default: 1e-6}

\item{rMotion}{double. Maximum x- and y distance a cell can travel by means of random motion per minute Default: 0.1 µm}
}
\value{
Object of class \code{growthSimulation}.
}
\description{
Method to initialise a \code{growthSimulation} object
}
\examples{
# Contruction a square environment of dimensions 300µm x 300µm x 5µm
sim <- init.simulation(cbind(c(-150, -150, 150, 150), c(-150, 150, 150, -150)),
                       gridFieldSize = 1.75, gridFieldLayers = 5)
}
