% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthSimulation.R
\name{plot.cells}
\alias{plot.cells}
\title{Plot spatial distribution of cells}
\usage{
plot.cells(object, xlim = NULL, ylim = NULL, ...)
}
\arguments{
\item{object}{S4-object of type \code{growthSimulation}.}

\item{xlim}{Numeric vector of length 2, specifying the x-range to be displayed.}

\item{ylim}{Numeric vector of length 2, specifying the y-range to be displayed.}

\item{iter}{Positive integer number of the simulation step/iteration to plot the cell distribution. If past simulation are displayed, the cell postitions needed to be recorded when running the simulation before (see \code{link{run.simulation}}). If NULL, current distribution is displayed.}
}
\value{
A ggplot object.
}
\description{
Plots the distribution of cells in a growth simulation.
}
