% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot.growth,growthSimulation-method}
\alias{plot.growth,growthSimulation-method}
\alias{plot.growth}
\title{Plot growth curves}
\usage{
\S4method{plot.growth}{growthSimulation}(object, tlim = NULL, ylim = NULL)
}
\arguments{
\item{object}{S4-object of class \link{growthSimulation}.}

\item{tlim}{Numeric vector of length 2, specifying the x-range (Time) to be displayed.}

\item{ylim}{Numeric vector of length 2, specifying the y-range (Mass) to be displayed.}
}
\value{
A \link{ggplot}
}
\description{
Plot population growth over time
}
